# argos-service-api/api.yml
...
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Argos Service
  license:
    name: Apache License 2.0
servers:
  - url: /api
security:
  - bearerAuth: []
paths:
...
  /supplychain:
    post:
      summary: create a supplychain
      operationId: createSupplyChain
      tags:
        - supplychain
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplyChain"
      responses:
        '201':
          description: created
          headers:
            Location:
              schema:
                type: string
                format: uri
              description: location where the created supplychain can be found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplyChain"
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/Error'
    get:
      summary: get supply chain by path
      operationId: getSupplyChainByPath
      tags:
        - supplychain
      parameters:
        - name: name
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/NameString"
        - name: path
          in: query
          style: simple
          explode: true
          required: true
          allowEmptyValue: false
          schema:
            type: array
            maxItems: 20
            items:
              $ref: "#/components/schemas/NameString"
      responses:
        '200':
          description: the supply chain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplyChain"
        '404':
          description: not found
        '500':
          $ref: '#/components/responses/Error'
  /supplychain/{supplyChainId}:
    get:
      summary: get a supplychain
      operationId: getSupplyChain
      tags:
        - supplychain
      parameters:
        - name: supplyChainId
          in: path
          description: supply chain id
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        '200':
          description: the supply chain item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplyChain"
        '500':
          $ref: '#/components/responses/Error'
    put:
      summary: update supplychain by id
      operationId: updateSupplyChain
      tags:
        - supplychain
      parameters:
        - name: supplyChainId
          in: path
          description: supply chain id
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplyChain"
      responses:
        '200':
          description: updated supply chain
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplyChain'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/Error'
    delete:
      summary: delete supply chain by id. The layout, approval configs and link objects are also deleted
      operationId: deleteSupplyChainById
      tags:
        - supplychain
      parameters:
        - name: supplyChainId
          in: path
          description: supply chain id
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        '204':
          description: supply chain deleted
        '500':
          $ref: '#/components/responses/Error'
...
  /supplychain/{supplyChainId}/link:
    get:
      summary: find link
      operationId: findLink
      tags:
        - link
      parameters:
        - name: supplyChainId
          in: path
          description: supply chain id
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
        - name: hash
          in: query
          description: hash of product or material
          required: false
          schema:
            $ref: "#/components/schemas/SHA256Hash"
      responses:
        '200':
          description: array of links
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LinkMetaBlock"
        '500':
          $ref: '#/components/responses/Error'
    post:
      summary: Create a link
      operationId: createLink
      tags:
        - link
      security:
        - basicAuth: []
        - bearerAuth: []
      parameters:
        - name: supplyChainId
          in: path
          description: supply chain id
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkMetaBlock"
      responses:
        '204':
          description: no content
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    basicAuth:
      type: http
      scheme: basic
  responses:
    ValidationError:
      description: validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationError"
    Error:
      description: server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    HexString:
      type: string
      maxLength: 1024
      pattern: '^[0-9a-f]*$'
    SHA256Hash:
      type: string
      minLength: 64
      maxLength: 64
      pattern: '^[0-9a-f]*$'
    UUID:
      type: string
      description: an Argos Service identifier
      minLength: 36
      maxLength: 36
      pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}?$'
    UnixPath:
      type: string
      maxLength: 1024
      pattern: '^(?!.*\\).*$'
    NameString:
      type: string
      minLength: 1
      maxLength: 63
      pattern: '^([a-z]|[a-z][a-z0-9-]*[a-z0-9])?$'
...
    ArtifactArray:
      type: array
      maxItems: 12288
      items:
        $ref: "#/components/schemas/Artifact"
...
    LinkMetaBlock:
      type: object
      required:
        - signature
        - link
      properties:
        signature:
          $ref: "#/components/schemas/Signature"
        link:
          $ref: "#/components/schemas/Link"
    Signature:
      type: object
      required:
        - keyId
        - signature
        - hashAlgorithm
        - keyAlgorithm
      properties:
        keyId:
          $ref: '#/components/schemas/KeyId'
        signature:
          $ref: "#/components/schemas/HexString"
        hashAlgorithm:
          $ref: '#/components/schemas/HashAlgorithm'
        keyAlgorithm:
          $ref: '#/components/schemas/KeyAlgorithm'
    Link:
      type: object
      required:
        - runId
        - layoutSegmentName
        - stepName
        - materials
        - products
      properties:
        runId:
          type: string
          maxLength: 1024
        stepName:
          $ref: "#/components/schemas/NameString"
        layoutSegmentName:
          $ref: "#/components/schemas/NameString"
        materials:
          $ref: "#/components/schemas/ArtifactArray"
        products:
          $ref: "#/components/schemas/ArtifactArray"
    Artifact:
      type: object
      required:
        - uri
        - hash
      properties:
        uri:
          $ref: "#/components/schemas/UnixPath"
        hash:
          $ref: "#/components/schemas/SHA256Hash"
...
    SupplyChain:
      type: object
      required:
        - name
        - parentLabelId
      properties:
        id:
          $ref: "#/components/schemas/UUID"
          readOnly: true
        name:
          $ref: "#/components/schemas/NameString"
        parentLabelId:
          $ref: "#/components/schemas/UUID"
...
    KeyId:
      $ref: "#/components/schemas/SHA256Hash"
    KeyAlgorithm:
      enum:
        - EC
    HashAlgorithm:
      enum:
        - SHA256
        - SHA384
        - SHA512
...
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    ValidationError:
      type: object
      required:
        - messages
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ValidationMessage"
    ValidationMessage:
      type: object
      required:
        - message
        - type
      properties:
        field:
          type: string
        type:
          enum:
            - DATA_INPUT
            - MODEL_CONSISTENCY
        message:
          type: string
...